refactor: InputChangeFunction Type으로 중복 제거
